refactor: extract useStatus hook
