+++ web/src/modules/invoices/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './get-invoice-number-from-id'
 export * from './parse-status'
+export * from './use-status'
